requirements.txt
Django==3.0.3
django-cors-headers==3.7.0
django-rest-swagger==2.2.0
djangorestframework==3.11.0
PyJWT==2.3.0